Fix `ExactOptionalize` changing `never` to `undefined` (#35)
`never` is a subtype of all types, so in this case, if `Schema[K]` is
never, `K` is used:

```ts
Schema[K] extends ExactOptionalStruct<any, any> ? K : never;
```

This is problematic when using a struct type like `Record<string,
never>`. This was previously incorrectly changed into `Record<string,
undefined>`. After this change, the type is not affected.